requirements.txt
feedparser>=6.0.0
pyyaml>=6.0
requests>=2.28.0
python-dateutil>=2.8.0
